+++ alu-web/src/lookup/lookup.ts
@@ -740,11 +740,13 @@ export async function apiReviewInstanceStudy(
   deckId: number,
   section: string,
   studyAhead: boolean,
+  essentialOnly: boolean,
 ): Promise<ReviewInstance[]> {
   return backendFetch<ReviewInstance[]>('POST', `decks/reviewinstance/study/`, {
     deck_id: deckId,
     section: section,
     study_ahead: studyAhead,
+    essential_only: essentialOnly,
   });
 }